% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSurvey_all_data.R
\name{SelectSurvey_all_data}
\alias{SelectSurvey_all_data}
\title{SelectSurvey_all_data}
\usage{
SelectSurvey_all_data()
}
\value{
Imports the data and makes separate Overview results for each question so that you can chose if you want to include all of the optional answears or not.
}
\description{
Running the command will ask you to select your tab delimited exported results from SelectSurvey and the Overview copy pasted into an excel file. Include your user start and end response dates to get response data. This function will import the data and  separate the Overview results for each question so that you can chose if you want to include all of the optional answears or not.
}
\examples{
SelectSurvey_all_data()
}
