% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSurvey_all_data.R
\name{SelectSurvey_exported_data}
\alias{SelectSurvey_exported_data}
\title{SelectSurvey_exported_data}
\usage{
SelectSurvey_exported_data()
}
\value{
Running the command will ask you to select your tab delimited exported results from SelectSurvey. Include your user start and end response dates to get response data.
}
\description{
SelectSurvey_exported_data
}
\examples{
SelectSurvey_exported_data()
}
