% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSov_plot.R
\name{SSov_plot}
\alias{SSov_plot}
\title{SSov_plot}
\usage{
SSov_plot(qn)
}
\arguments{
\item{qn}{Name of the object with the question from the overview to plot.}
}
\value{
A plot of the results from the overview for that question.
}
\description{
SSov_plot
}
\examples{
SSov_plot(Q_1)
}
